+++ app/jobs/prune_sessions_job.rb
@@ -1,0 +1,7 @@
+class PruneSessionsJob < ApplicationJob
+  queue_as :default
+
+  def perform
+    Session.expired.in_batches(of: 500).delete_all
+  end
+end