+++ test/controllers/concerns/authentication_test.rb
@@ -27,6 +27,7 @@ class AuthenticationTest < ActionDispatch::IntegrationTest
 
     assert session.reload.last_seen_at > 1.minute.ago
     assert_equal "Current Browser", session.user_agent
+    assert_equal session.last_seen_at, user.reload.last_seen_at
   end
 
   test "#update_session_activity should not touch recent session" do
@@ -43,16 +44,18 @@ class AuthenticationTest < ActionDispatch::IntegrationTest
     assert_equal old_timestamp.to_i, session.reload.last_seen_at.to_i
   end
 
-  test "#update_session_activity should establish a newly issued session" do
+  test "#terminate_session should preserve the user last seen after logout" do
     user = create(:user)
     sign_in_as(user)
     session = user.sessions.last
 
-    assert_nil session.last_seen_at
+    last_seen_at = session.last_seen_at
+    assert_equal last_seen_at, user.reload.last_seen_at
 
-    get feeds_path
+    delete session_path
 
-    assert session.reload.last_seen_at > 1.minute.ago
+    assert_not Session.exists?(session.id)
+    assert_equal last_seen_at, user.reload.last_seen_at
   end
 
   test "#require_authentication should reject a session past the inactivity timeout" do